+++ docs/app/pages/components/sections/media-player/media-player/media-player.component.html
@@ -80,7 +80,7 @@
     <tab heading="HTML">
         <uxd-snippet [content]="snippets.compiled.appHtml"></uxd-snippet>
     </tab>
-    <tab heading="TypesScript">
+    <tab heading="TypeScript">
         <uxd-snippet [content]="snippets.compiled.appTs"></uxd-snippet>
     </tab>
     <tab heading="CSS">
