+++ docs/app/pages/components/sections/media-player/media-player/snippets/app.ts
@@ -2,7 +2,8 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'app',
-    templateUrl: './src/app.component.html'
+    templateUrl: './src/app.component.html',
+    styleUrls: ['./src/app.component.css']
 })
 export class AppComponent {
