+++ src/components/media-player/extensions/controls/controls.component.less
@@ -12,18 +12,18 @@ ux-media-player-controls {
         transition: color 0.3s ease-in-out;
 
         &:hover {
-            color: @grey2;
+            color: @mediaplayer-controls-btn-hover-color;
         }
     }
 
     .control-button {
-        fill: @grey3;
+        fill: @mediaplayer-controls-btn-color;
         margin: 0 10px 8px;
         cursor: pointer;
         transition: fill 0.3s ease-in-out;
 
         &:hover {
-            fill: @grey2;
+            fill: @mediaplayer-controls-btn-hover-color;
         }
     }
 
@@ -33,33 +33,33 @@ ux-media-player-controls {
 
     &.quiet {
         .hpe-icon {
-            color: @grey4;
+            color: @mediaplayer-controls-quiet-color;
 
             &:hover {
-                color: @grey6;
+                color: @mediaplayer-controls-quiet-hover-color;
             }
         }
 
         .control-button {
-            fill: @grey4;
+            fill: @mediaplayer-controls-quiet-color;
 
             &:hover {
-                fill: @grey6;
+                fill: @mediaplayer-controls-quiet-hover-color;
             }
         }
 
         .volume-container {
             .volume-slider-container {
                 .volume-slider {
-                    background-color: rgba(207, 207, 207, 0.1);
+                    background-color: @mediaplayer-volume-slider-quiet-track-upper;
 
                     .volume-track-lower {
-                        background-color: rgba(153, 153, 153, 0.5);
+                        background-color: @mediaplayer-volume-slider-quiet-track-lower;
                     }
 
                     .volume-slider-thumb {
-                        background-color: @white;
-                        border: 1px solid @grey3;
+                        background-color: @mediaplayer-volume-slider-quiet-thumb-bg;
+                        border: 1px solid @mediaplayer-volume-slider-quiet-thumb-border;
                     }
                 }
             }
@@ -90,12 +90,12 @@ ux-media-player-controls {
                 position: relative;
                 width: 70px;
                 height: 4px;
-                background-color: @grey6;
+                background-color: @mediaplayer-volume-slider-track-upper;
                 margin: 10px 7px;
                 transition: opacity 0.3s ease-in-out;
 
                 .volume-track-lower {
-                    background-color: @grey3;
+                    background-color: @mediaplayer-volume-slider-track-lower;
                     height: 100%;
                 }
 
@@ -104,8 +104,8 @@ ux-media-player-controls {
                     width: 14px;
                     height: 14px;
                     border-radius: 7px;
-                    background-color: @white;
-                    border: 1px solid @grey3;
+                    background-color: @mediaplayer-volume-slider-thumb-bg;
+                    border: 1px solid @mediaplayer-volume-slider-thumb-border;
                     top: -5px;
                     cursor: pointer;
                     transform: translateX(-50%);
