+++ src/components/media-player/media-player.component.html
@@ -1,14 +1,14 @@
 <div class="video-player-container" *ngIf="type === 'video'">
 
-    <video class="video-player" #player [src]="source" (abort)="abortEvent.next()" (canplay)="canPlayEvent.next()"
-        (canplaythrough)="canPlayThroughEvent.next()" (durationchange)="durationChangeEvent.next(player.duration)" (ended)="endedEvent.next()"
-        (error)="errorEvent.next($event)" (loadeddata)="loadedDataEvent.next($event)" (loadedmetadata)="loadedMetadataEvent.next($event)"
-        (loadstart)="loadStartEvent.next()" (pause)="pauseEvent.next()" (play)="playEvent.next()" (playing)="playingEvent.next(!player.paused)"
-        (ratechange)="rateChangeEvent.next(player.playbackRate)" (seeked)="seekedEvent.next(player.currentTime)" (seeking)="seekingEvent.next(player.currentTime)"
-        (stalled)="stalledEvent.next()" (suspend)="suspendEvent.next()" (timeupdate)="timeUpdateEvent.next(player.currentTime)"
-        (volumechange)="volumeChangeEvent.next(player.volume)" (waiting)="waitingEvent.next()" (click)="mediaClickEvent.next($event)"></video>
+    <video class="video-player" #player [src]="source" (abort)="mediaPlayerService.abortEvent.next()" (canplay)="mediaPlayerService.canPlayEvent.next()"
+        (canplaythrough)="mediaPlayerService.canPlayThroughEvent.next()" (durationchange)="mediaPlayerService.durationChangeEvent.next(player.duration)" (ended)="mediaPlayerService.endedEvent.next()"
+        (error)="mediaPlayerService.errorEvent.next($event)" (loadeddata)="mediaPlayerService.loadedDataEvent.next($event)" (loadedmetadata)="mediaPlayerService.loadedMetadataEvent.next($event)"
+        (loadstart)="mediaPlayerService.loadStartEvent.next()" (pause)="mediaPlayerService.pauseEvent.next()" (play)="mediaPlayerService.playEvent.next()" (playing)="mediaPlayerService.playingEvent.next(!player.paused)"
+        (ratechange)="mediaPlayerService.rateChangeEvent.next(player.playbackRate)" (seeked)="mediaPlayerService.seekedEvent.next(player.currentTime)" (seeking)="mediaPlayerService.seekingEvent.next(player.currentTime)"
+        (stalled)="mediaPlayerService.stalledEvent.next()" (suspend)="mediaPlayerService.suspendEvent.next()" (timeupdate)="mediaPlayerService.timeUpdateEvent.next(player.currentTime)"
+        (volumechange)="mediaPlayerService.volumeChangeEvent.next(player.volume)" (waiting)="mediaPlayerService.waitingEvent.next()" (click)="mediaPlayerService.mediaClickEvent.next($event)"></video>
 
-    <div class="video-overlay" [class.playing]="playing | async">
+    <div class="video-overlay" [class.playing]="mediaPlayerService.playing | async">
         <svg class="play-graphic" x="0px" y="0px" viewBox="0 0 64 64">
             <circle class="play-circle" cx="32.2" cy="31.8" r="31.8" />
             <polygon class="play-triangle" points="23,14.1 23,50.8 48.3,32.5" />
@@ -40,13 +40,13 @@
     <p class="audio-file-format">{{ (audioMetadata | async)?.description }}</p>
     <p class="audio-file-size">{{ (audioMetadata | async)?.size | fileSize }}</p>
 
-    <audio #player [src]="source" (abort)="abortEvent.next()" (canplay)="canPlayEvent.next(true)"
-        (canplaythrough)="canPlayThroughEvent.next(true)" (durationchange)="durationChangeEvent.next(player.duration)" (ended)="endedEvent.next()"
-        (error)="errorEvent.next($event)" (loadeddata)="loadedDataEvent.next($event)" (loadedmetadata)="loadedMetadataEvent.next($event)"
-        (loadstart)="loadStartEvent.next()" (pause)="pauseEvent.next()" (play)="playEvent.next()" (playing)="playingEvent.next(!player.paused)"
-        (ratechange)="rateChangeEvent.next(player.playbackRate)" (seeked)="seekedEvent.next(player.currentTime)" (seeking)="seekingEvent.next(player.currentTime)"
-        (stalled)="stalledEvent.next()" (suspend)="suspendEvent.next()" (timeupdate)="timeUpdateEvent.next(player.currentTime)"
-        (volumechange)="volumeChangeEvent.next(player.volume)" (waiting)="waitingEvent.next()" (click)="mediaClickEvent.next($event)"></audio>
+    <audio #player [src]="source" (abort)="mediaPlayerService.abortEvent.next()" (canplay)="mediaPlayerService.canPlayEvent.next(true)"
+        (canplaythrough)="mediaPlayerService.canPlayThroughEvent.next(true)" (durationchange)="mediaPlayerService.durationChangeEvent.next(player.duration)" (ended)="mediaPlayerService.endedEvent.next()"
+        (error)="mediaPlayerService.errorEvent.next($event)" (loadeddata)="mediaPlayerService.loadedDataEvent.next($event)" (loadedmetadata)="mediaPlayerService.loadedMetadataEvent.next($event)"
+        (loadstart)="mediaPlayerService.loadStartEvent.next()" (pause)="mediaPlayerService.pauseEvent.next()" (play)="mediaPlayerService.playEvent.next()" (playing)="mediaPlayerService.playingEvent.next(!player.paused)"
+        (ratechange)="mediaPlayerService.rateChangeEvent.next(player.playbackRate)" (seeked)="mediaPlayerService.seekedEvent.next(player.currentTime)" (seeking)="mediaPlayerService.seekingEvent.next(player.currentTime)"
+        (stalled)="mediaPlayerService.stalledEvent.next()" (suspend)="mediaPlayerService.suspendEvent.next()" (timeupdate)="mediaPlayerService.timeUpdateEvent.next(player.currentTime)"
+        (volumechange)="mediaPlayerService.volumeChangeEvent.next(player.volume)" (waiting)="mediaPlayerService.waitingEvent.next()" (click)="mediaPlayerService.mediaClickEvent.next($event)"></audio>
 </div>
 
 <div class="control-bar">
