+++ src/components/media-player/media-player.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input, ViewChild, ElementRef, AfterViewInit, OnDestroy } from '@angular/core';
 import { ExtractedFrame, FrameExtractionService } from '../../services/frame-extraction/frame-extraction.service';
+import { MediaPlayerService } from './media-player.service';
 import { Subject } from 'rxjs/Subject';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { Observable } from 'rxjs/Observable';
@@ -12,394 +13,90 @@ import 'rxjs/add/operator/switchMap';
 @Component({
     selector: 'ux-media-player',
     templateUrl: './media-player.component.html',
+    providers: [ MediaPlayerService ],
     host: {
         'tabindex': '0',
-        '(keydown.Space)': 'togglePlay()',
-        '[class.standard]': '!fullscreen',
-        '[class.fullscreen]': 'fullscreen',
-        '[class.quiet]': 'quietMode && type === "video" || fullscreen',
+        '(keydown.Space)': 'mediaPlayerService.togglePlay()',
+        '[class.standard]': '!mediaPlayerService.fullscreen',
+        '[class.fullscreen]': 'mediaPlayerService.fullscreen',
+        '[class.quiet]': 'quietMode && type === "video" || mediaPlayerService.fullscreen',
         '[class.hover]': 'hovering',
         '[class.video]': 'type === "video"',
         '[class.audio]': 'type === "audio"',
         '(mouseenter)': 'hovering = true',
         '(mouseleave)': 'hovering = false',
-        '(document:webkitfullscreenchange)': 'fullscreenChange($event)',
-        '(document:mozfullscreenchange)': 'fullscreenChange($event)',
-        '(document:MSFullscreenChange)': 'fullscreenChange($event)'
+        '(document:webkitfullscreenchange)': 'mediaPlayerService.fullscreenChange($event)',
+        '(document:mozfullscreenchange)': 'mediaPlayerService.fullscreenChange($event)',
+        '(document:MSFullscreenChange)': 'mediaPlayerService.fullscreenChange($event)'
     }
 })
 export class MediaPlayerComponent implements AfterViewInit, OnDestroy {
-
-    @Input() source: string;
-    @Input() type: MediaPlayerType = 'video';
 
     @ViewChild('player') private _playerRef: ElementRef;
     @ViewChild('trackBar') private _trackBarRef: ElementRef;
 
     hovering: boolean = false;
     audioMetadata: Observable<AudioMetadata>;
-    playing: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
+    
+    private _hover$: Subscription;
+    private _clicked$: Subscription;
+    private _paused$: Subscription;
+    private _playing$: Subscription;
+    private _loading$: Subscription;
 
-    /*
-        Create observables for media player events
-    */
-    initEvent: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
-    abortEvent: Subject<void> = new Subject<void>();
-    canPlayEvent: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
-    canPlayThroughEvent: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
-    durationChangeEvent: Subject<number> = new Subject<number>();
-    endedEvent: Subject<void> = new Subject<void>();
-    errorEvent: Subject<any> = new Subject<any>();
-    loadedDataEvent: Subject<any> = new Subject<any>();
-    loadedMetadataEvent: Subject<any> = new Subject<any>();
-    loadStartEvent: Subject<void> = new Subject<void>();
-    pauseEvent: Subject<void> = new Subject<void>();
-    playEvent: Subject<void> = new Subject<void>();
-    playingEvent: Subject<boolean> = new Subject<boolean>();
-    rateChangeEvent: Subject<number> = new Subject<number>();
-    seekedEvent: Subject<number> = new Subject<number>();
-    seekingEvent: Subject<number> = new Subject<number>();
-    stalledEvent: Subject<void> = new Subject<void>();
-    suspendEvent: Subject<void> = new Subject<void>();
-    timeUpdateEvent: Subject<number> = new Subject<number>();
-    volumeChangeEvent: Subject<number> = new Subject<number>();
-    waitingEvent: Subject<void> = new Subject<void>();
-    mediaClickEvent: Subject<MouseEvent> = new Subject<MouseEvent>();
-    fullscreenEvent: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
-    quietModeEvent: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
-    progressEvent: Observable<TimeRanges> = Observable.create((observer: Observer<TimeRanges>) => {
+    get source(): string {
+        return this.mediaPlayerService.source;
+    }
 
-        // repeat until the whole video has fully loaded
-        let interval = setInterval(() => {
+    @Input() 
+    set source(value: string) {
+        this.mediaPlayerService.source = value;
+    }
 
-            let buffered = this._playerRef.nativeElement.buffered as TimeRanges;
-            observer.next(buffered);
+    get type(): MediaPlayerType {
+        return this.mediaPlayerService.type;
+    }
 
-            if (buffered.length === 1 && buffered.start(0) === 0 && buffered.end(0) === this.duration) {
-                observer.complete();
-                clearInterval(interval);
-            }
-        }, 1000);
-    });
-
-    private _mediaPlayer: HTMLMediaElement;
-    private _fullscreen: boolean = false;
-    private _quietMode: boolean = false;
-    private _mousemoveSub: Subscription;
-    private _clickSub: Subscription;
-    private _pauseSub: Subscription;
-    private _playingSub: Subscription;
-    private _loaded: boolean = false;
-    private _loadedSub: Subscription;
+    @Input() 
+    set type(value: MediaPlayerType) {
+        this.mediaPlayerService.type = value;
+    }    
 
     get quietMode(): boolean {
-        return this._quietMode;
+        return this.mediaPlayerService.quietMode;
     }
 
     @Input()
     set quietMode(value: boolean) {
-
-        // quiet mode cannot be enabled on audio player
-        if (this.type === 'audio') {
-            value = false;
-        }
-
-        this._quietMode = value;
-        this.quietModeEvent.next(value);
+        this.mediaPlayerService.quietMode = value;
     }
 
-    /*
-        Create all the getters and setters the can be used by media player extensions 
-    */
-    get mediaPlayer(): HTMLMediaElement {
-        return this._mediaPlayer;
-    }
+    constructor(public mediaPlayerService: MediaPlayerService, private _audioService: AudioService, private _elementRef: ElementRef) {
 
-    get mediaPlayerWidth(): number {
-        return this._mediaPlayer ? this._mediaPlayer.offsetWidth : 0;
-    }
-
-    get mediaPlayerHeight(): number {
-        return this._mediaPlayer ? this._mediaPlayer.offsetHeight : 0;
-    }
-
-    get audioTracks(): AudioTrackList {
-        return this._mediaPlayer ? this._mediaPlayer.audioTracks : null;
-    }
-
-    get autoplay(): boolean {
-        return this._mediaPlayer ? this._mediaPlayer.autoplay : false;
-    }
-    set autoplay(value: boolean) {
-        this._mediaPlayer.autoplay = value;
-    }
-
-    get buffered(): TimeRanges {
-        return this._mediaPlayer ? this._mediaPlayer.buffered : new TimeRanges();
-    }
-
-    get crossOrigin(): string {
-        return this._mediaPlayer ? this._mediaPlayer.crossOrigin : null;
-    }
-    set crossOrigin(value: string) {
-        this._mediaPlayer.crossOrigin = value;
-    }
-
-    get currentSrc(): string {
-        return this._mediaPlayer ? this._mediaPlayer.currentSrc : null;
-    }
-
-    get currentTime(): number {
-        return this._mediaPlayer ? this._mediaPlayer.currentTime : 0;
-    }
-    set currentTime(value: number) {
-        this._mediaPlayer.currentTime = value;
-    }
-
-    get defaultMuted(): boolean {
-        return this._mediaPlayer ? this._mediaPlayer.defaultMuted : false;
-    }
-    set defaultMuted(value: boolean) {
-        this._mediaPlayer.defaultMuted = value;
-    }
-
-    get defaultPlaybackRate(): number {
-        return this._mediaPlayer ? this._mediaPlayer.defaultPlaybackRate : 1;
-    }
-    set defaultPlaybackRate(value: number) {
-        this._mediaPlayer.defaultPlaybackRate = value;
-    }
-
-    get duration(): number {
-        return this._mediaPlayer ? this._mediaPlayer.duration : 0;
-    }
-
-    get ended(): boolean {
-        return this._mediaPlayer ? this._mediaPlayer.ended : false;
-    }
-
-    get loop(): boolean {
-        return this._mediaPlayer ? this._mediaPlayer.loop : false;
-    }
-    set loop(value: boolean) {
-        this._mediaPlayer.loop = value;
-    }
-
-    get muted(): boolean {
-        return this._mediaPlayer ? this._mediaPlayer.muted : false;
-    }
-    set muted(value: boolean) {
-        this._mediaPlayer.muted = value;
-    }
-
-    get networkState(): number {
-        return this._mediaPlayer.networkState;
-    }
-
-    get paused(): boolean {
-        return this._mediaPlayer ? this._mediaPlayer.paused : true;
-    }
-
-    get playbackRate(): number {
-        return this._mediaPlayer ? this._mediaPlayer.playbackRate : 1;
-    }
-    set playbackRate(value: number) {
-        this._mediaPlayer.playbackRate = value;
-    }
-
-    get played(): TimeRanges {
-        return this._mediaPlayer ? this._mediaPlayer.played : new TimeRanges();
-    }
-
-    get preload(): string {
-        return this._mediaPlayer ? this._mediaPlayer.preload : 'auto';
-    }
-    set preload(value: string) {
-        this._mediaPlayer.preload = value;
-    }
-
-    get readyState(): number {
-        return this._mediaPlayer ? this._mediaPlayer.readyState : 0;
-    }
-
-    get seekable(): TimeRanges {
-        return this._mediaPlayer ? this._mediaPlayer.seekable : new TimeRanges();
-    }
-
-    get seeking(): boolean {
-        return this._mediaPlayer ? this._mediaPlayer.seeking : false;
-    }
-
-    get src(): string {
-        return this._mediaPlayer ? this._mediaPlayer.src : '';
-    }
-    set src(value: string) {
-        this._mediaPlayer.src = value;
-    }
-
-    get textTracks(): TextTrackList {
-        return this._mediaPlayer ? this._mediaPlayer.textTracks : new TextTrackList();
-    }
-
-    get videoTracks(): VideoTrackList {
-        return this._mediaPlayer ? this._mediaPlayer.videoTracks : new VideoTrackList();
-    }
-
-    get volume(): number {
-        return this._mediaPlayer ? this._mediaPlayer.volume : 1;
-    }
-    set volume(value: number) {
-        this._mediaPlayer.volume = value;
-    }
-
-    get fullscreen(): boolean {
-        return this._mediaPlayer ? this._fullscreen : false;
-    }
-    set fullscreen(value: boolean) {
-        this._fullscreen = value;
-        this.fullscreenEvent.next(value);
-    }
-
-    constructor(private _frameExtractionService: FrameExtractionService, private _audioService: AudioService, private _elementRef: ElementRef) {
-
-        this._mousemoveSub = Observable.fromEvent(this._elementRef.nativeElement, 'mousemove').switchMap((event: MouseEvent) => {
+        // show controls when hovering and in quiet mode
+        this._hover$ = Observable.fromEvent(this._elementRef.nativeElement, 'mousemove').switchMap((event: MouseEvent) => {
             this.hovering = true;         
             return Observable.of(event);
         }).debounceTime(2000).subscribe(() => this.hovering = false);
     }
 
     ngAfterViewInit(): void {
-        this._mediaPlayer = this._playerRef.nativeElement as HTMLMediaElement;
+        this.mediaPlayerService.setMediaPlayer(this._elementRef.nativeElement, this._playerRef.nativeElement);
 
-        this.audioMetadata = this._audioService.getAudioFileMetadata(this._mediaPlayer);
-        this._playingSub = this.playingEvent.subscribe(event => this.playing.next(true));
-        this._pauseSub = this.pauseEvent.subscribe(event => this.playing.next(false));
-        this._clickSub = this.mediaClickEvent.subscribe(() => this.togglePlay());
-        this._loadedSub = this.loadedMetadataEvent.subscribe(() => this._loaded = true);
-
-        this.initEvent.next(true);
+        this.audioMetadata = this._audioService.getAudioFileMetadata(this._playerRef.nativeElement);
+        this._playing$ = this.mediaPlayerService.playingEvent.subscribe(event => this.mediaPlayerService.playing.next(true));
+        this._paused$ = this.mediaPlayerService.pauseEvent.subscribe(event => this.mediaPlayerService.playing.next(false));
+        this._clicked$ = this.mediaPlayerService.mediaClickEvent.subscribe(() => this.mediaPlayerService.togglePlay());
+        this._loading$ = this.mediaPlayerService.loadedMetadataEvent.subscribe(() => this.mediaPlayerService.loaded = true);
     }
 
     ngOnDestroy(): void {
-        this._mousemoveSub.unsubscribe();
-        this._playingSub.unsubscribe();
-        this._pauseSub.unsubscribe();
-        this._clickSub.unsubscribe();
-        this._loadedSub.unsubscribe();
+        this._hover$.unsubscribe();
+        this._playing$.unsubscribe();
+        this._paused$.unsubscribe();
+        this._clicked$.unsubscribe();
+        this._loading$.unsubscribe();
     }
-
-    /**
-     * Starts playing the audio/video
-     */
-    play(): void {
-        this._mediaPlayer.play();
-    }
-
-    /**
-     * Pauses the currently playing audio/video
-     */
-    pause(): void {
-        this._mediaPlayer.pause();
-    }
-
-    /**
-     * Toggle playing state
-     */
-    togglePlay(): void {
-
-        // prevent any action is not loaded
-        if (this._loaded === false) {
-            return;
-        }
-
-        if (this.paused) {
-            this.play();
-        } else {
-            this.pause();
-        }
-    }
-
-    /**
-     * Re-loads the audio/video element
-     */
-    load(): void {
-        this._mediaPlayer.load();
-    }
-
-    /**
-     * Checks if the browser can play the specified audio/video type
-     */
-    canPlayType(type: string): string {
-        return this._mediaPlayer.canPlayType(type);
-    }
-
-    /**
-     * Adds a new text track to the audio/video
-     */
-    addTextTrack(kind: string, label: string, language: string): TextTrack {
-        return this._mediaPlayer.addTextTrack(kind, label, language);
-    }
-
-    /**
-     * Attempt to display media in fullscreen mode
-     */
-    requestFullscreen(): void {
-
-        if (this._elementRef.nativeElement.requestFullscreen) {
-            this._elementRef.nativeElement.requestFullscreen();
-        } else if (this._elementRef.nativeElement.webkitRequestFullscreen) {
-            this._elementRef.nativeElement.webkitRequestFullscreen();
-        } else if ((<any>this._elementRef.nativeElement).msRequestFullscreen) {
-            (<any>this._elementRef.nativeElement).msRequestFullscreen();
-        } else if ((<any>this._elementRef.nativeElement).mozRequestFullScreen) {
-            (<any>this._elementRef.nativeElement).mozRequestFullScreen();
-        }
-    }
-
-    /**
-     * Exit full screen mode
-     */
-    exitFullscreen(): void {
-
-        if (this._elementRef.nativeElement.exitFullscreen) {
-            document.exitFullscreen();
-        } else if (document.webkitExitFullscreen) {
-            document.webkitExitFullscreen();
-        } else if ((<any>document).msExitFullscreen) {
-            (<any>document).msExitFullscreen();
-        } else if ((<any>document).mozCancelFullScreen) {
-            (<any>document).mozCancelFullScreen();
-        }
-    }
-
-    fullscreenChange(event: Event) {
-        this.fullscreen = (<any>document).fullscreen || document.webkitIsFullScreen || (<any>document).mozFullScreen || (<any>document).msFullscreenElement !== null && (<any>document).msFullscreenElement !== undefined;
-        this.fullscreenEvent.next(this.fullscreen);
-    }
-
-    /**
-     * Toggle Fullscreen State
-     */
-    toggleFullscreen(): void {
-        if (this.fullscreen) {
-            this.exitFullscreen();
-        } else {
-            this.requestFullscreen();
-        }
-    }
-
-    /**
-     * Extract the frames from the video
-     */
-    getFrames(width: number, height: number, skip: number): Observable<ExtractedFrame> {
-
-        if (this.type === 'video') {
-            return this._frameExtractionService.getFrameThumbnails(this.source, width, height, 0, this.duration, 10);
-        }
-
-        return Observable.from([]);
-    }
-
 }
 
 export type MediaPlayerType = 'video' | 'audio';
