+++ src/styles/variables.less
@@ -1247,3 +1247,37 @@
 @page-header-nav-dropdown-hover-bg: @page-header-nav-active-bg;
 @page-header-nav-dropdown-hover-color: @grey6;
 @page-header-nav-dropdown-selected-color: @white;
+
+// Media Player
+@mediaplayer-shadow: 0 1px 1px 2px @grey6;
+@mediaplayer-play-btn-bg: rgba(0, 0, 0, 0.5);
+@mediaplayer-play-btn-color: rgba(255, 255, 255, 0.4);
+@mediaplayer-controls-bg: @white;
+@mediaplayer-controls-quiet-bg: rgba(0, 0, 0, 0.7);
+@mediaplayer-controls-quiet-color: @grey4;
+@mediaplayer-controls-quiet-hover-color: @grey6;
+@mediaplayer-controls-btn-color: @grey3;
+@mediaplayer-controls-btn-hover-color: @grey2;
+@mediaplayer-audio-bg: @white;
+@mediaplayer-audio-border: @grey8;
+@mediaplayer-audio-shadow: inset 0 -2px 3px -1px @grey6;
+@mediaplyer-audio-title: @grey3;
+@mediaplyer-audio-subtitle: @grey4;
+@mediaplayer-volume-slider-track-upper: @grey6;
+@mediaplayer-volume-slider-track-lower: @grey3;
+@mediaplayer-volume-slider-thumb-bg: @white;
+@mediaplayer-volume-slider-thumb-border: @grey3;
+@mediaplayer-volume-slider-quiet-track-upper: rgba(207, 207, 207, 0.1);
+@mediaplayer-volume-slider-quiet-track-lower: rgba(153, 153, 153, 0.5);
+@mediaplayer-volume-slider-quiet-thumb-bg: @white;
+@mediaplayer-volume-slider-quiet-thumb-border: @grey3;
+@mediaplayer-timeline-bg: @grey6;
+@mediaplayer-timeline-buffered-bg: @grey5;
+@mediaplayer-timeline-track: @primary;
+@mediaplayer-timeline-handle: @grey2;
+@mediaplayer-timeline-text: @grey3;
+@mediaplayer-timeline-quiet-bg: rgba(232, 232, 232, 0.2);
+@mediaplayer-timeline-quiet-buffered-bg: rgba(207, 207, 207, 0.1);
+@mediaplayer-timeline-quiet-track: @primary;
+@mediaplayer-timeline-quiet-handle: @grey4;
+@mediaplayer-timeline-quiet-text: @grey4;
